+++ lbh15/properties/common_interface.py
@@ -27,3 +27,33 @@ class OxygenDiffusivityInterface(PropertyInterface):
         str : Oxygen diffusivity long name
         """
         return "oxygen diffusivity"
+
+
+class OxygenPartialPressureInterface(PropertyInterface):
+    """
+    *Oxygen partial pressure in liquid metal divided by the
+    Oxygen concentration in liquid metal squared* property abstract class.
+    """
+    @property
+    def name(self) -> str:
+        """
+        str : Name of the property
+        """
+        return "o_pp"
+
+    @property
+    def units(self) -> str:
+        """
+        str : Oxygen partial pressure in liquid metal divided by the
+        Oxygen concentration in liquid metal squared unit
+        """
+        return "[Pa.wt.%^-2]"
+
+    @property
+    def long_name(self) -> str:
+        """
+        str : Oxygen partial pressure in liquid metal divided by the
+        Oxygen concentration in liquid metal squared long name
+        """
+        return ("Oxygen partial pressure divided by the"
+                " oxygen concentration squared")
