+++ lbh15/properties/lead_thermochemical_properties/lead_thermochemical.py
@@ -6,6 +6,7 @@ import numpy as np
 from scipy.constants import atm
 from scipy.constants import R
 from lbh15.properties.interface import PropertyInterface
+from lbh15.properties.common_interface import OxygenPartialPressureInterface
 from ..lead_properties import h
 from ..._commons import LEAD_MELTING_TEMPERATURE as T_m0
 from ..._commons import LEAD_MOLAR_MASS as M
@@ -13,7 +14,7 @@ from ..._commons import OXYGEN_MOLAR_MASS as M_O
 from ..._decorators import range_warning
 
 
-class OxygenPartialPressureInterface(PropertyInterface):
+class LeadOxygenPartialPressureInterface(OxygenPartialPressureInterface):
     """
     *Oxygen partial pressure in liquid lead divided by the
     Oxygen concentration in liquid lead squared* property abstract class.
@@ -41,30 +42,6 @@ class OxygenPartialPressureInterface(PropertyInterface):
         return 1600
 
     @property
-    def name(self) -> str:
-        """
-        str : Name of the property
-        """
-        return "o_pp"
-
-    @property
-    def units(self) -> str:
-        """
-        str : Oxygen partial pressure in liquid lead divided by the
-        Oxygen concentration in liquid lead squared unit
-        """
-        return "[Pa.wt.%^-2]"
-
-    @property
-    def long_name(self) -> str:
-        """
-        str : Oxygen partial pressure in liquid lead divided by the
-        Oxygen concentration in liquid lead squared long name
-        """
-        return ("Oxygen partial pressure divided by the"
-                " oxygen concentration squared")
-
-    @property
     def description(self) -> str:
         """
         str : Oxygen partial pressure in liquid lead divided by the
@@ -73,7 +50,7 @@ class OxygenPartialPressureInterface(PropertyInterface):
         return f"{self.long_name} in liquid lead"
 
 
-class OxygenPartialPressureOtsuka1979(OxygenPartialPressureInterface):
+class OxygenPartialPressureOtsuka1979(LeadOxygenPartialPressureInterface):
     """
     *Oxygen partial pressure in liquid lead divided by the
     Oxygen concentration in liquid lead squared* property class
@@ -124,7 +101,7 @@ class OxygenPartialPressureOtsuka1979(OxygenPartialPressureInterface):
         return [1073, 1673]
 
 
-class OxygenPartialPressureOtsuka1981(OxygenPartialPressureInterface):
+class OxygenPartialPressureOtsuka1981(LeadOxygenPartialPressureInterface):
     """
     *Oxygen partial pressure in liquid lead divided by the
     Oxygen concentration in liquid lead squared* property class
@@ -175,7 +152,7 @@ class OxygenPartialPressureOtsuka1981(OxygenPartialPressureInterface):
         return [1023, 1273]
 
 
-class OxygenPartialPressureGanesan2006(OxygenPartialPressureInterface):
+class OxygenPartialPressureGanesan2006(LeadOxygenPartialPressureInterface):
     """
     *Oxygen partial pressure in liquid lead divided by the
     Oxygen concentration in liquid lead squared* property class
@@ -226,7 +203,7 @@ class OxygenPartialPressureGanesan2006(OxygenPartialPressureInterface):
         return [815, 1090]
 
 
-class OxygenPartialPressureAlcock1964(OxygenPartialPressureInterface):
+class OxygenPartialPressureAlcock1964(LeadOxygenPartialPressureInterface):
     """
     *Oxygen partial pressure in liquid lead divided by the
     Oxygen concentration in liquid lead squared* property class
@@ -277,7 +254,7 @@ class OxygenPartialPressureAlcock1964(OxygenPartialPressureInterface):
         return [783, 973]
 
 
-class OxygenPartialPressureSzwarc1972(OxygenPartialPressureInterface):
+class OxygenPartialPressureSzwarc1972(LeadOxygenPartialPressureInterface):
     """
     *Oxygen partial pressure in liquid lead divided by the
     Oxygen concentration in liquid lead squared* property class
@@ -328,7 +305,7 @@ class OxygenPartialPressureSzwarc1972(OxygenPartialPressureInterface):
         return [1012, 1353]
 
 
-class OxygenPartialPressureCharle1976(OxygenPartialPressureInterface):
+class OxygenPartialPressureCharle1976(LeadOxygenPartialPressureInterface):
     """
     *Oxygen partial pressure in liquid lead divided by the
     Oxygen concentration in liquid lead squared* property class
@@ -379,7 +356,7 @@ class OxygenPartialPressureCharle1976(OxygenPartialPressureInterface):
         return [1173, 1373]
 
 
-class OxygenPartialPressureIsecke1977(OxygenPartialPressureInterface):
+class OxygenPartialPressureIsecke1977(LeadOxygenPartialPressureInterface):
     """
     *Oxygen partial pressure in liquid lead divided by the
     Oxygen concentration in liquid lead squared* property class
@@ -430,7 +407,7 @@ class OxygenPartialPressureIsecke1977(OxygenPartialPressureInterface):
         return [1173, 1373]
 
 
-class OxygenPartialPressureTaskinen1979(OxygenPartialPressureInterface):
+class OxygenPartialPressureTaskinen1979(LeadOxygenPartialPressureInterface):
     """
     *Oxygen partial pressure in liquid lead divided by the
     Oxygen concentration in liquid lead squared* property class
@@ -481,7 +458,7 @@ class OxygenPartialPressureTaskinen1979(OxygenPartialPressureInterface):
         return [1073, 1203]
 
 
-class OxygenPartialPressureFisher1966(OxygenPartialPressureInterface):
+class OxygenPartialPressureFisher1966(LeadOxygenPartialPressureInterface):
     """
     *Oxygen partial pressure in liquid lead divided by the
     Oxygen concentration in liquid lead squared* property class
